+++ src/python/pants/base/exceptions.py
@@ -3,8 +3,6 @@
 
 from __future__ import annotations
 
-from typing import Iterable
-
 
 class TargetDefinitionException(Exception):
     """Indicates an invalid target definition.
@@ -30,17 +28,3 @@ class BackendConfigurationError(BuildConfigurationError):
 
 class MappingError(Exception):
     """Indicates an error mapping addressable objects."""
-
-
-class ResolveError(MappingError):
-    """Indicates an error resolving targets."""
-
-    @classmethod
-    def did_you_mean(
-        cls, *, bad_name: str, known_names: Iterable[str], namespace: str
-    ) -> ResolveError:
-        possibilities = "\n  ".join(f":{target_name}" for target_name in sorted(known_names))
-        return cls(
-            f"'{bad_name}' was not found in namespace '{namespace}'. Did you mean one "
-            f"of:\n  {possibilities}"
-        )
