+++ src/python/pants/build_graph/address.py
@@ -7,8 +7,9 @@ import dataclasses
 import os
 from dataclasses import dataclass
 from pathlib import PurePath
-from typing import Any, Mapping, Sequence
+from typing import Any, Iterable, Mapping, Sequence
 
+from pants.base.exceptions import MappingError
 from pants.engine.engine_aware import EngineAwareParameter
 from pants.engine.internals import native_engine
 from pants.engine.internals.native_engine import (  # noqa: F401
@@ -623,3 +624,17 @@ class BuildFileAddress:
 
     address: Address
     rel_path: str
+
+
+class ResolveError(MappingError):
+    """Indicates an error resolving target addresses."""
+
+    @classmethod
+    def did_you_mean(
+        cls, *, bad_name: str, known_names: Iterable[str], namespace: str
+    ) -> ResolveError:
+        possibilities = "\n  ".join(f":{target_name}" for target_name in sorted(known_names))
+        return cls(
+            f"'{bad_name}' was not found in namespace '{namespace}'. Did you mean one "
+            f"of:\n  {possibilities}"
+        )
